requirements.txt
streamlit>=1.50.0
plotly>=6.0.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
